inciando o desenvolvimento de in memory cache

+++ src/main/java/com/project/fileservice/Service/FileService.java
@@ -1,10 +1,7 @@
 package com.project.fileservice.Service;
 
 import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.ListObjectsRequest;
-import com.amazonaws.services.s3.model.ObjectMetadata;
-import com.amazonaws.services.s3.model.S3Object;
-import com.amazonaws.services.s3.model.S3ObjectSummary;
+import com.amazonaws.services.s3.model.*;
 import com.project.fileservice.Exception.InvalidFileException;
 import com.project.fileservice.Exception.S3ConnectionError;
 import com.project.fileservice.Model.DownloadDTO;
@@ -15,7 +12,9 @@ import org.springframework.stereotype.Service;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Slf4j
@@ -24,6 +23,7 @@ public class FileService {
     private String bucketName;
     private final AmazonS3 s3client;
     private final ListObjectsRequest listS3Objects;
+    private final List<S3ObjectSummary> s3ObjectSummaries = new ArrayList<>();
 
     public FileService(AmazonS3 amazonS3,ListObjectsRequest listObjectsRequest){
         this.s3client=amazonS3;
@@ -32,9 +32,15 @@ public class FileService {
     public List<S3ObjectSummary> listBucketItems(){
         log.info("Listando objetos da bucket....");
         try{
-            return s3client
+            List<S3ObjectSummary> objectSummaryList = s3client
                     .listObjects(listS3Objects)
                     .getObjectSummaries();
+            if(this.s3ObjectSummaries.isEmpty()){
+                s3ObjectSummaries.addAll(objectSummaryList);
+                return objectSummaryList;
+            }
+            log.info("Pegando do in memory cache");
+            return objectSummaryList;
         }catch (Exception e){
             throw new S3ConnectionError("Nao foi possivel se conectar a AWS e retornar a lista de objetos");
         }
@@ -67,7 +73,12 @@ public class FileService {
     public void deleteItem(String key){
         log.info("Deletando item da bucket...");
         try{
-            s3client.deleteObject(bucketName,key);
+            this.s3ObjectSummaries.forEach(s3ObjectSummary -> {
+                if(s3ObjectSummary.getKey().equals(key)){
+                    s3client.deleteObject(bucketName,key);
+                    this.s3ObjectSummaries.remove(s3ObjectSummary);
+                }
+            });
         }catch (Exception e){
             throw new S3ConnectionError("Nao foi possivel se conectar com a AWS e deletar o item");
         }
